requirements.txt
langchain
langchain-ollama
langchain-openai
langchain_groq
langchain-community
langchain-text-splitters
faiss-cpu
chromadb
python-dotenv
tiktoken
pypdf
fastapi
uvicorn